// packages/shared-cli-rest/src/helpers/templates/types.ts
export interface OperationInfo {
  /**
   * Contains the operation ID, which is required.
   */
  operationId: string;
  /**
   * Contains the function type, which will be imported to the controller.
   */
  functionType: string;
  /**
   * Contains the parameters type for the operation. This is so that in the controller-types file,
   * the generator can import the right parameter.
   */
  parametersName?: string;
  /**
   * Contains the variable names of the response object schema.
   */
  responseType: {
    success: string;
    error?: string;
  };
}